ml_service/requirements.txt
numpy>=1.24.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.9
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0